base non-ui files moved to core module
