--- app/src/main/java/lostankit7/droid/moodtracker/base/entity/Status.kt
@@ -1,7 +1,0 @@
-package lostankit7.droid.moodtracker.base.entity
-
-sealed class Status<T> {
-    class Success<T> : Status<T>()
-    object Loading : Status<Nothing>()
-    data class Error(val message: String) : Status<Nothing>()
-}
