+++ app/src/main/java/lostankit7/droid/moodtracker/domain/Mapper.kt
@@ -3,7 +3,7 @@ package lostankit7.droid.moodtracker.domain
 import lostankit7.droid.moodtracker.data.database.entities.TaskIcon
 import lostankit7.droid.moodtracker.data.database.entities.UserEntry
 import lostankit7.droid.moodtracker.model.MoodEntry
-import lostankit7.droid.moodtracker.presentation.fragment.splash.Constants
+import lostankit7.droid.moodtracker.core.utils.Constants
 
 object Mapper {
     fun mapDataToUserEntry(
