+++ app/src/main/java/lostankit7/droid/moodtracker/presentation/fragment/splash/SplashFragment.kt
@@ -7,6 +7,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.delay
 import lostankit7.droid.moodtracker.base.fragment.BaseDaggerFragment
+import lostankit7.droid.moodtracker.core.utils.Constants
 import lostankit7.droid.moodtracker.databinding.FragmentSplashBinding
 import lostankit7.droid.moodtracker.di.AppComponent
 import lostankit7.droid.moodtracker.utils.constant.SplashFragmentDataProvider
