+++ app/src/main/java/lostankit7/droid/moodtracker/presentation/viewmodel/MoodEntryViewModel.kt
@@ -3,7 +3,6 @@ package lostankit7.droid.moodtracker.presentation.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import lostankit7.droid.moodtracker.R
-import lostankit7.droid.moodtracker.base.entity.StringHandler
 import lostankit7.droid.moodtracker.base.viewmodel.BaseViewModel
 import lostankit7.droid.moodtracker.data.database.entities.MoodIcon
 import lostankit7.droid.moodtracker.data.repository.MoodIconRepository
@@ -17,8 +16,8 @@ class MoodEntryViewModel @Inject constructor(
     private val suggestedMoodNameRepository: SuggestedMoodNameRepository,
 ) : BaseViewModel() {
 
-    private val _errorMessageLiveData = MutableLiveData<StringHandler>()
-    val errorMessageLiveData: LiveData<StringHandler> get() = _errorMessageLiveData
+    private val _errorMessageLiveData = MutableLiveData<lostankit7.droid.moodtracker.core.utils.entity.UiText>()
+    val errorMessageLiveData: LiveData<lostankit7.droid.moodtracker.core.utils.entity.UiText> get() = _errorMessageLiveData
 
     val suggestedMoodNamesLiveData get() = suggestedMoodNameRepository.suggestedMoodNames
     val moodIconsLiveData: LiveData<List<MoodIcon>> get() = moodRepository.moodIcons
@@ -27,7 +26,7 @@ class MoodEntryViewModel @Inject constructor(
     fun saveMoodIcon(moodName: String, moodIcon: String, iconId: Int?) {
         when {
             moodName.isBlank() -> {
-                _errorMessageLiveData.value = StringHandler.ResourceString(R.string.enter_mood_name)
+                _errorMessageLiveData.value = lostankit7.droid.moodtracker.core.utils.entity.UiText.ResourceString(R.string.enter_mood_name)
             }
             else -> {
                 val icon = MoodIcon(moodIcon, moodName)
