+++ core/src/main/java/lostankit7/droid/moodtracker/core/di/viewmodel/ViewModelKey.kt
@@ -1,4 +1,4 @@
-package lostankit7.droid.moodtracker.di
+package lostankit7.droid.moodtracker.core.di.viewmodel
 
 import androidx.lifecycle.ViewModel
 import dagger.MapKey
@@ -8,4 +8,4 @@ import kotlin.reflect.KClass
 @Target(AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.RUNTIME)
 @MapKey
-internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
+annotation class ViewModelKey(val value: KClass<out ViewModel>)
