+++ core/src/main/java/lostankit7/droid/moodtracker/core/utils/entity/UiEvent.kt
@@ -1,0 +1,7 @@
+package lostankit7.droid.moodtracker.core.utils.entity
+
+sealed class UiEvent<T> {
+    class Success<T> : UiEvent<T>()
+    object Loading : UiEvent<Nothing>()
+    data class Error(val message: String) : UiEvent<Nothing>()
+}
